% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStarvMort.R
\name{getStarvMort}
\alias{getStarvMort}
\alias{starvMort}
\title{Get starvation mortality}
\usage{
getStarvMort(
  params,
  n = initialN(params),
  n_pp = initialNResource(params),
  n_other = initialNOther(params),
  t = 0,
  ...
)

starvMort(params, n, n_pp, n_other, t = 0, ...)
}
\arguments{
\item{params}{A \link[mizer:MizerParams]{MizerParams} object}

\item{n}{A matrix of species abundances (species x size).}

\item{n_pp}{A vector of the plankton abundance by size}

\item{n_other}{A list of abundances for other dynamical components of the
ecosystem}

\item{t}{The time for which to do the calculation. Defaults to 0.}

\item{...}{Unused}
}
\value{
A two dimensional array of instantaneous starvation mortality
(species x size).
}
\description{
There is no starvation mortality as long as the energy income rate
\eqn{E_r} is positive. For details of this rate see
\code{mizer::getEReproAndGrowth()}. Once this rate is negative, the per-capita
mortality is proportional to this rate and inversely proportional to body
weight (and therefore also lipid reserves):
\deqn{\mu_s(w) = \frac{-E_r(w)}{w} {\tt starv\_coeff} }{mu_s(w) = -E_r(w)/w * starv_coeff}
The proportionality constant \code{starv_coeff} is set with \code{setStarvation()}.
}
\concept{rate functions}
