% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStarvMort.R
\name{getStarvMort}
\alias{getStarvMort}
\title{Get starvation mortality}
\usage{
getStarvMort(
  params,
  n = params@initial_n,
  n_pp = params@initial_n_pp,
  n_other = params@initial_n_other,
  ...
)
}
\arguments{
\item{params}{A \linkS4class{MizerParams} object}

\item{n}{A matrix of species abundances (species x size).}

\item{n_pp}{A vector of the plankton abundance by size}

\item{n_other}{A list of abundances for other dynamical components of the
ecosystem}

\item{...}{Unused}
}
\value{
A two dimensional array of instantaneous starvation mortality
(species x size).
}
\description{
Starvation mortality is proportional to the energy deficiency and
is
There is no starvation mortality as long as the energy income rate
\eqn{E_r} is positive. For details of this rate see
\code{mizer::getEReproAndGrowth()}. Once this rate is negative, the per-capita
mortality is proportional to this rate and inversely proportional to body
weight (and therefore also lipid reserves):
\deqn{\mu_s(w) = \frac{E_r(w)}{w} {\tt starv_coeff} }{mu_s(w) = E_r(w)/w * starv\_coeff}
The proportionality constant is set with \code{setStarvation()}.
}
\concept{rate functions}
