% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{mizerStarvation}
\alias{mizerStarvation}
\alias{mizerStarvation-package}
\title{mizerStarvation: Starvation mortality in mizer}
\description{
This is an extension package for the mizer package
(https://sizespectrum.org/mizer/) to implement starvation mortality.
}
