% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setStarvation.R
\name{setStarvation}
\alias{setStarvation}
\title{Set starvation mortality}
\usage{
setStarvation(params, starv_coef = 10)
}
\arguments{
\item{params}{A MizerParams object}

\item{starv_coef}{Proportionality constant for starvation mortality. When
\code{starv_coef} is equal to 10 the instantaneous starvation mortality (1/year)
is 1 when energy deficit is 10\% of body weight. When \code{starv_coef = 0} there
is no starvation mortality}
}
\value{
A MizerParams object with starvation mortality
}
\description{
Set starvation mortality
}
